fix(#194): apply review findings to the biome terrain layer
- widen the Worley scan to 5x5 so nearest and second-nearest feature
  points are always found, and dedupe the two scan helpers into one
- fold build-biome-sample into get-biome, keeping a single copy of the
  permanent forbidden-pattern comment
- guard buildBiomeField's resolution as a positive integer
- key the ground tints off BIOME_ROSTER so a grown roster fails loudly
- test the modifier draw path, same-size pan byte rewrites, and the
  wobble channels by name; reorder supporting declarations to first use
- cover buildBiomeField's viewport parameter in the worldmap doc
